adds cumulative statistics orders program selection according to application start preselects most recent program renames component to "statistics"
