+++ frontend-nx/apps/edu-hub/i18n.json
@@ -9,7 +9,7 @@
       "course-admin-page",
       "course-application",
       "course-page",
-      "enrollmentStatistics",
+      "statistics",
       "manageAchievementTemplates",
       "manageAppSettings",
       "manageCourse",
