+++ frontend-nx/apps/edu-hub/locales/de/statistics.json
@@ -3,5 +3,7 @@
     "select_programs": {
         "label": "Progamme oder Semester",
         "placeholder": "Progamm oder Semester hinzufügen"
-    }
+    },
+    "daily_enrollments": "Tägliche Bewerbungen",
+    "cumulative_enrollments": "Kumulierte Bewerbungen"
 }
