--- frontend-nx/apps/edu-hub/locales/en/enrollmentStatistics.json
@@ -1,7 +1,0 @@
-{
-    "enrollment_statistics": "Enrollment Statistics",
-    "select_programs": {
-        "label": "Programs or Semesters",
-        "placeholder": "Add Program or Semester"
-    }
-}
